Rework OSC packet classes
classes to contain updated API to replace older API which is deprecated and contained inside abstract packet classes (prefixed with A)

+++ src/main/java/oscP5/OscArgument.java
@@ -1,87 +1,75 @@
-/**
- * An OSC (Open Sound Control) library for processing.
- *
- * ##copyright##
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- * 
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- * 
- * You should have received a copy of the GNU Lesser General
- * Public License along with this library; if not, write to the
- * Free Software Foundation, Inc., 59 Temple Place, Suite 330,
- * Boston, MA  02111-1307  USA
- * 
- * @author		##author##
- * @modified	##date##
- * @version		##version##
- */
-
 package oscP5;
 
 public final class OscArgument {
 
 	private Object value;
 
-	OscArgument() {
+	OscArgument()
+	{
 	}
 
-	public int intValue() {
+	public int intValue()
+	{
 		return ((Integer) value).intValue();
 	}
 
-	public char charValue() {
+	public char charValue()
+	{
 		return ((Character) value).charValue();
 	}
 
-	public float floatValue() {
+	public float floatValue()
+	{
 		return ((Float) value).floatValue();
 	}
 
-	public double doubleValue() {
+	public double doubleValue()
+	{
 		return ((Double) value).doubleValue();
 	}
 
-	public long longValue() {
+	public long longValue()
+	{
 		return ((Long) value).longValue();
 	}
 
-	public boolean booleanValue() {
+	public boolean booleanValue()
+	{
 		return ((Boolean) value).booleanValue();
 	}
 
-	public String stringValue() {
+	public String stringValue()
+	{
 		return ((String) value);
 	}
 
-	public String toString() {
+	public String toString()
+	{
 		return ((String) value);
 	}
 
-	public byte[] bytesValue() {
+	public byte[] bytesValue()
+	{
 		return ((byte[]) value);
 	}
 
-	public byte[] blobValue() {
+	public byte[] blobValue()
+	{
 		return ((byte[]) value);
 	}
 
-	public int[] midiValue() {
+	public int[] midiValue()
+	{
 		final int[] myInt = new int[4];
 		final byte[] myByte = (byte[]) value;
 		for (int i = 0; i < 4; i++) {
-			myInt[i] = (int) (myByte[i]);
+			myInt[i] = myByte[i];
 		}
 		return (myInt);
 	}
 
-	void setValue(final Object theValue) {
+	void setValue(final Object theValue)
+	{
 		value = theValue;
 	}
 }
